requirements.txt
acachecontrol==0.3.6
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-mgmt-compute==34.0.0
azure-mgmt-core==1.5.0
azure-mgmt-loganalytics==12.0.0
azure-mgmt-monitor==6.0.2
azure-mgmt-reservations==2.3.0
azure-monitor-query==1.4.1
boto3==1.36.4
botocore==1.36.4
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
concurrent-log-handler==0.9.28
cryptography==45.0.6
datadog==0.51.0
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.124.0
frozenlist==1.7.0
google-api-core==2.22.0
google-auth==2.36.0
google-cloud-billing==1.16.2
google-cloud-compute==1.20.1
google-cloud-core==2.4.1
google-cloud-monitoring==2.27.2
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.12.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
jmespath==1.0.1
lxml==6.0.0
MarkupSafe==3.0.2
motor==3.7.1
msal==1.33.0
msal-extensions==1.3.1
msoffcrypto-tool==5.4.2
msrest==0.7.1
multidict==6.6.4
numpy==2.0.2
oauthlib==3.3.1
okta-jwt-verifier==0.2.7
olefile==0.47
openai==1.97.0
openpyxl==3.1.5
pandas==2.2.3
pillow==11.3.0
portalocker==3.2.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
retry2==0.9.5
rsa==4.9.1
s3transfer==0.11.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.50.0
tabulate==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
xlsxwriter==3.2.5
yarl==1.20.1
